+++ tests/entrypoints/openai/test_embedding.py
@@ -1,9 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import base64
-from collections.abc import Sequence
 from http import HTTPStatus
-from typing import Optional
 
 import numpy as np
 import openai
@@ -14,15 +12,11 @@ from vllm.entrypoints.openai.protocol import EmbeddingResponse
 from vllm.transformers_utils.tokenizer import get_tokenizer
 
 from ...conftest import HfRunner
-from ...models.embedding.utils import (EmbedModelInfo, check_embeddings_close,
-                                       matryoshka_fy)
+from ...models.embedding.utils import EmbedModelInfo, correctness_test
 from ...utils import RemoteOpenAIServer
 
 MODELS = [
-    EmbedModelInfo("intfloat/multilingual-e5-small", is_matryoshka=False),
-    EmbedModelInfo("Snowflake/snowflake-arctic-embed-m-v1.5",
-                   is_matryoshka=True,
-                   matryoshka_dimensions=[256]),
+    EmbedModelInfo("intfloat/multilingual-e5-small", is_matryoshka=False)
 ]
 
 DUMMY_CHAT_TEMPLATE = """{% for message in messages %}{{message['role'] + ': ' + message['content'] + '\\n'}}{% endfor %}"""  # noqa: E501
@@ -56,14 +50,6 @@ def server(model_info, dtype: str):
         "--chat-template",
         DUMMY_CHAT_TEMPLATE
     ]
-
-    if model_info.name == "Snowflake/snowflake-arctic-embed-m-v1.5":
-        # Manually enable Matryoshka Embeddings
-        args.extend([
-            "--trust_remote_code", "--hf_overrides",
-            '{"matryoshka_dimensions":[256]}'
-        ])
-
     with RemoteOpenAIServer(model_info.name, args) as remote_server:
         yield remote_server
 
@@ -115,7 +101,7 @@ async def test_single_embedding(model_info: EmbedModelInfo,
     assert embeddings.usage.total_tokens > 0
 
     vllm_outputs = [d.embedding for d in embeddings.data]
-    _correctness_test(hf_model, prompts, vllm_outputs)
+    correctness_test(hf_model, prompts, vllm_outputs)
 
     # test using token IDs
     input_tokens = [1, 1, 1, 1, 1]
@@ -161,7 +147,7 @@ async def test_batch_embedding(model_info: EmbedModelInfo,
     assert embeddings.usage.total_tokens > 0
 
     vllm_outputs = [d.embedding for d in embeddings.data]
-    _correctness_test(hf_model, prompts, vllm_outputs)
+    correctness_test(hf_model, prompts, vllm_outputs)
 
     # test list[list[int]]
     input_tokens = [[4, 5, 7, 9, 20], [15, 29, 499], [24, 24, 24, 24, 24],
@@ -252,7 +238,7 @@ async def test_batch_base64_embedding(model_info: EmbedModelInfo,
                                                      model=model_info.name,
                                                      encoding_format="float")
     float_data = [d.embedding for d in responses_float.data]
-    _correctness_test(hf_model, prompts, float_data)
+    correctness_test(hf_model, prompts, float_data)
 
     # test base64 responses
     responses_base64 = await client.embeddings.create(input=prompts,
@@ -263,13 +249,13 @@ async def test_batch_base64_embedding(model_info: EmbedModelInfo,
         base64_data.append(
             np.frombuffer(base64.b64decode(data.embedding),
                           dtype="float32").tolist())
-    _correctness_test(hf_model, prompts, base64_data)
+    correctness_test(hf_model, prompts, base64_data)
 
     # Default response is float32 decoded from base64 by OpenAI Client
     responses_default = await client.embeddings.create(input=prompts,
                                                        model=model_info.name)
     default_data = [d.embedding for d in responses_default.data]
-    _correctness_test(hf_model, prompts, default_data)
+    correctness_test(hf_model, prompts, default_data)
 
 
 @pytest.mark.asyncio
@@ -325,77 +311,3 @@ async def test_embedding_truncation(
         assert "error" in response.object
         assert "truncate_prompt_tokens value is greater than max_model_len. "\
                "Please, select a smaller truncation size." in response.message
-
-
-@pytest.mark.asyncio
-async def test_matryoshka(model_info: EmbedModelInfo,
-                          server: RemoteOpenAIServer, hf_model: HfRunner):
-    client = server.get_async_client()
-
-    async def make_request_and_correctness_test(dimensions):
-        prompts = input_texts * 3
-
-        embedding_response = await client.embeddings.create(
-            model=model_info.name,
-            input=prompts,
-            dimensions=dimensions,
-            encoding_format="float",
-        )
-        embeddings = EmbeddingResponse.model_validate(
-            embedding_response.model_dump(mode="json"))
-
-        assert embeddings.id is not None
-        assert len(embeddings.data) == 3
-        assert len(embeddings.data[0].embedding) > 0
-        assert embeddings.usage.completion_tokens == 0
-        assert embeddings.usage.prompt_tokens > 0
-        assert embeddings.usage.total_tokens > 0
-
-        if dimensions is not None:
-            assert len(embeddings.data[0].embedding) == dimensions
-
-        vllm_outputs = [d.embedding for d in embeddings.data]
-        _correctness_test(hf_model, prompts, vllm_outputs, dimensions)
-
-    if model_info.is_matryoshka:
-        valid_dimensions: list[Optional[int]] = [None]
-        if model_info.matryoshka_dimensions is not None:
-            valid_dimensions += model_info.matryoshka_dimensions[:2]
-
-        for dimensions in valid_dimensions:
-            await make_request_and_correctness_test(dimensions)
-
-        invalid_dimensions: list[Optional[int]] = [-1]
-        if model_info.matryoshka_dimensions is not None:
-            assert 5 not in model_info.matryoshka_dimensions
-            invalid_dimensions.append(5)
-
-        for dimensions in invalid_dimensions:
-            with pytest.raises(openai.BadRequestError):
-                await make_request_and_correctness_test(dimensions)
-
-    else:
-        for dimensions in [None]:
-            await make_request_and_correctness_test(dimensions)
-
-        for dimensions in [-1, 16]:
-            with pytest.raises(openai.BadRequestError):
-                await make_request_and_correctness_test(dimensions)
-
-
-def _correctness_test(hf_model: HfRunner,
-                      inputs,
-                      vllm_outputs: Sequence[list[float]],
-                      dimensions: Optional[int] = None):
-
-    hf_outputs = hf_model.encode(inputs)
-    if dimensions:
-        hf_outputs = matryoshka_fy(hf_outputs, dimensions)
-
-    check_embeddings_close(
-        embeddings_0_lst=hf_outputs,
-        embeddings_1_lst=vllm_outputs,
-        name_0="hf",
-        name_1="vllm",
-        tol=1e-2,
-    )
